refactor: move access_result_to_error to github_access module
Move access_result_to_error and access_error_to_app_error from
routes/read/access.rs to github_access/error_conversion.rs to
eliminate the service -> routes layer violation.

Also change pub mod services to mod services (crate-internal only).

Closes review comment on #96.
