+++ crates/api/src/lib.rs
@@ -6,6 +6,7 @@ pub mod github_access;
 pub mod middleware;
 pub mod pagination;
 pub mod routes;
+mod services;
 
 use std::sync::Arc;
