+++ crates/api/src/routes/read/board_projects.rs
@@ -9,11 +9,11 @@ use crate::extractors::AuthenticatedSession;
 use crate::github_access::DynGithubAccessChecker;
 use crate::pagination::{PaginatedResponse, PaginationParams, encode_cursor};
 
-use super::access::access_result_to_error;
 use super::dto::{
     BoardProjectDetailResponse, BoardProjectListItem, RepositoryRef, derive_board_project_state,
     parse_board_run_status,
 };
+use crate::github_access::access_result_to_error;
 
 // ─── GET /api/v1/repositories/{github_repository_id}/board-projects ──────────
