+++ crates/api/src/routes/read/mod.rs
@@ -1,4 +1,3 @@
-mod access;
 mod artifacts;
 mod board_projects;
 mod board_runs;
@@ -8,7 +7,6 @@ mod findings;
 mod repositories;
 mod viewer_sources;
 
-pub use access::*;
 pub use artifacts::*;
 pub use board_projects::*;
 pub use board_runs::*;
